Move timeline blueprint to own class
